Remove duplicated issues options and some more refactors
